Fix race condition
If the timer is reset by a new event, but then, before the timer
triggers a run, an older event comes in, the timer is reset again and
lastChange is now the old timestamp. If the timer triggers now, the
old timestamp is seen and no run is triggered. In other words: With
bad luck, the run is missed.

This fixes the behavior by only resetting the timer when a newer
event than the last run comes in. If the timer eventually triggers,
the run is executed in any case.

Signed-off-by: beorn7 <beorn@grafana.com>